pyproject.toml
[tool.poetry]
name = "controlecolaboradoresapi"
version = "0.1.0"
description = ""
authors = ["Hélder Lima Gonçalves de Oliveira <helderlgoliveira@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.7"
djangorestframework = "^3.12.4"
drf-access-policy = "^1.1.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.1"
argon2-cffi = "^21.1.0"

[tool.poetry.dev-dependencies]
coverage = "^5.5"
pytest = "^6.2.4"
pytest-django = "^4.4.0"
pytest-mock = "^3.6.1"
model-bakery = "^1.3.2"
pytest-cov = "^2.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
